updated containers

+++ containers/synapse/scripts/python/postprocess_cpu.py
@@ -16,36 +16,33 @@ from n5_utils import read_n5_block, write_n5_block
 
 def tif_read(file_name):
     """
-    read tif image in (rows,cols,slices) shape
+    read tif image as (slices, rows, cols) shape
+    and returns it as (cols, rows, slices
+    so the addressing of the in memory array will be [x, y, z]
     """
     im = skimage.io.imread(file_name)
-    im_array = np.zeros(
-        (im.shape[1], im.shape[2], im.shape[0]), dtype=np.uint16)
-    for i in range(im.shape[0]):
-        im_array[:, :, i] = im[i]
+    im_array = im.transpose(2, 1, 0)
     return im_array
 
 
 def tif_write(im_array, file_name):
     """
-    write an array with (rows,cols,slices) shape into a tif image
+    the input im_array has the shape (cols, rows, slices)
+    and it writes it as (slices, rows, cols)
     """
-    im = np.zeros((im_array.shape[2], im_array.shape[0],
-                   im_array.shape[1]), dtype=im_array.dtype)
-    for i in range(im_array.shape[2]):
-        im[i] = im_array[:, :, i]
+    im = im_array.transpose(2, 1, 0)
     skimage.io.imsave(file_name, im)
     return None
 
 
-def remove_small_piece(out_path, prefix, img, start, end, mask=None, threshold=10, percentage=1.0):
+def remove_small_piece(out_path, prefix, img, start, end, mask=None, threshold=10, percentage=1.0, connectivity=3):
     """
     Remove blobs that have less than N voxels.
     Write final result to output hdf5 file, output a .csv file indicating the location and size of each synapses.
     Args:
     out_path: location to write CSV files
     prefix: prefix for CSV filenames
-    img: image to operate on
+    img: image to operate on in the shape (cols, rows, slices)
     start: (x,y,z) tuple indicating starting corner
     end: (x,y,z) tuple indicating ending corner
     mask: optional mask image
@@ -55,7 +52,7 @@ def remove_small_piece(out_path, prefix, img, start, end, mask=None, threshold=1
 
     print("Removing small blobs and save results to disk...")
     img[img != 0] = 1
-    label_img = label(img, connectivity=3)
+    label_img = label(img, connectivity=connectivity)
     regionprop_img = regionprops(label_img)
     idx = 0
 
@@ -70,7 +67,10 @@ def remove_small_piece(out_path, prefix, img, start, end, mask=None, threshold=1
         print("num voxels: ", num_voxel)
         curr_obj = np.zeros(img.shape, dtype=img.dtype)
         curr_obj[label_img == props.label] = 1
-        center_row, center_col, center_vol = props.centroid
+        # because the image array has shape (cols, rows, slices)
+        # the coord comming from region properties will be
+        # in the form (x, y, z) instead of (slice, row, col)
+        center_x, center_y, center_z = props.centroid
 
         print("  Non zero: ", np.count_nonzero(curr_obj))
 
@@ -79,16 +79,17 @@ def remove_small_piece(out_path, prefix, img, start, end, mask=None, threshold=1
         else:
             mask = np.ones(img.shape, dtype=img.dtype)
         curr_obj = curr_obj * mask
-        print("  Non zero after masking: ", np.count_nonzero(curr_obj))
+        num_masked_voxels = np.count_nonzero(curr_obj)
+        print("  Non zero after masking: ", num_masked_voxels)
 
         exclude = False
-        if num_voxel < threshold:
+        if num_masked_voxels < threshold:
             exclude = True
         if percentage < 1:
-            if np.count_nonzero(curr_obj) < num_voxel*percentage:
+            if num_masked_voxels < num_voxel*percentage:
                 exclude = True
         else:
-            if mask[int(center_row), int(center_col), int(center_vol)] == 0:
+            if mask[int(center_x), int(center_y), int(center_z)] == 0:
                 exclude = True
 
         if exclude:
@@ -102,20 +103,20 @@ def remove_small_piece(out_path, prefix, img, start, end, mask=None, threshold=1
                     writer = csv.writer(
                         csv_file, delimiter=',', quotechar='"', quoting=csv.QUOTE_MINIMAL)
                     writer.writerow(
-                        [' ID ', ' Num vxl ', ' centroid ', ' bbox row ', ' bbox col ', ' bbox vol '])
+                        [' ID ', ' Num vxl ', ' centroid ', ' bbox x ', ' bbox y ', ' bbox vol '])
 
             idx += 1
-            min_row, min_col, min_vol, max_row, max_col, max_vol = props.bbox
-            bbox_row = (int(min_row+start[0]), int(max_row+start[0]))
-            bbox_col = (int(min_col+start[1]), int(max_col+start[1]))
-            bbox_vol = (int(min_vol+start[2]), int(max_vol+start[2]))
+            min_x, min_y, min_z, max_x, max_y, max_z = props.bbox
+            bbox_x = (int(min_x+start[0]), int(max_x+start[0]))
+            bbox_y = (int(min_y+start[1]), int(max_y+start[1]))
+            bbox_z = (int(min_z+start[2]), int(max_z+start[2]))
 
-            center = (int(center_row+start[0]),
-                      int(center_col+start[1]),
-                      int(center_vol+start[2]))
+            center = (int(center_x+start[0]),
+                      int(center_y+start[1]),
+                      int(center_z+start[2]))
 
             csv_row = [str(idx), str(num_voxel), str(center),
-                       str(bbox_row), str(bbox_col), str(bbox_vol)]
+                       str(bbox_x), str(bbox_y), str(bbox_z)]
             with open(csv_filepath, 'a') as csv_file:
                 writer = csv.writer(csv_file, delimiter=',',
                                     quotechar='"', quoting=csv.QUOTE_MINIMAL)
@@ -159,6 +160,11 @@ def main():
 
     parser.add_argument('-p', '--percentage', dest='percentage', type=float, default=1.0,
                         help='threshold to remove the object if it falls in the mask less than a percentage. If percentage is 1, criteria will be whether the centroid falls within the mask.')
+
+    parser.add_argument('-c', '--connectivity', dest='connectivity', type=int, default=3,
+                        help='Specify region connectivity')
+    parser.add_argument('--keep_ws_tiff', dest='keep_ws_tiff', action='store_true', default=False,
+                        help='If true keep the tiffs generated by the watershed')
 
     args = parser.parse_args()
     start = tuple([int(d) for d in args.start_coord.split(',')])
@@ -215,11 +221,12 @@ def main():
                              end,
                              mask=mask,
                              threshold=args.threshold,
-                             percentage=args.percentage)
+                             percentage=args.percentage,
+                             connectivity=args.connectivity)
 
     write_n5_block(args.output_path, args.data_set, start, end, img)
 
-    if os.path.exists(out_img_path):
+    if not args.keep_ws_tiff and os.path.exists(out_img_path):
         os.remove(out_img_path)
 
     elapsed_time = time.time() - start_time
